requirement.txt
asgiref==3.5.1
Django==3.2
django-ckeditor==6.4.0
django-js-asset==2.0.0
mysqlclient==2.1.0
Pillow==9.1.0
pytz==2022.1
sqlparse==0.4.2
